Deleted public-messages and made public methods inside MessagesResource accessible without authentication Started fixing log messages
