+++ src/main/java/com/app/storytel/challenge/resources/MessagesResource.java
@@ -38,7 +38,6 @@ public class MessagesResource {
 
     @PostMapping
     public ResponseEntity<?> createNewMessage(@Valid @RequestBody MessageRequest messageRequest) {
-        log.info("Call made to saveNewMessage method");
         ErrorMessage errorMessage;
         Message message = this.messageService.saveNewMessage(messageRequest,
                 loginInformationComponent.fetchLoginUserInformation());
@@ -47,10 +46,10 @@ public class MessagesResource {
             MessageResponse messageResponse = new MessageResponse(message.getId(), message.getSubject(),
                     message.getMessageContent(), message.getViews(), message.getOwner().getId(), message.getCreated(),
                     message.getModified());
-            log.info("Call to saveNewMessage was successful");
+            log.debug("Call to saveNewMessage was successful");
             return new ResponseEntity<>(messageResponse, HttpStatus.CREATED);
         } else {
-            log.info("Failed to save new message information for {}", messageRequest);
+            log.error("Failed to save new message information for {}", messageRequest);
             errorMessage = new ErrorMessage("Create failed", new Date());
         }
 
@@ -59,21 +58,20 @@ public class MessagesResource {
 
     @PutMapping
     public ResponseEntity<?> updateExistingMessage(@Valid @RequestBody MessageRequest messageRequest) {
-        log.info("Call made to updateExistingMessage method");
         ErrorMessage errorMessage;
 
         if (messageRequest.getViews() != null) {
             if (this.messageService.updateMessage(messageRequest,
                     loginInformationComponent.fetchLoginUserInformation())) {
 
-                log.info("Call to updateExistingMessage was successful");
+                log.debug("Call to updateExistingMessage was successful");
                 Message updatedMessage = this.messageService.findMessage(messageRequest.getId());
                 MessageResponse messageResponse = new MessageResponse(updatedMessage.getId(), updatedMessage.getSubject(),
                         updatedMessage.getMessageContent(), updatedMessage.getViews(), updatedMessage.getOwner().getId(),
                         updatedMessage.getCreated(), updatedMessage.getModified());
                 return new ResponseEntity<>(messageResponse, HttpStatus.OK);
             } else {
-                log.info("Failed to update message with record: {}", messageRequest);
+                log.error("Failed to update message with record: {}", messageRequest);
                 errorMessage = new ErrorMessage("Bad Request", new Date());
             }
         } else {
@@ -91,15 +89,54 @@ public class MessagesResource {
 
     @DeleteMapping(path = "/{messageId}")
     public ResponseEntity<Void> deleteMessage(@PathVariable Long messageId) {
-        log.info("Call made to deleteMessage method");
 
         if (this.messageService.deleteMessage(messageId, loginInformationComponent.fetchLoginUserInformation())) {
-            log.info("Message with id {} deleted", messageId);
             return ResponseEntity.noContent().build();
         } else {
-            log.info("Failed to delete message");
+            log.error("Failed to delete message");
         }
 
         return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+    }
+
+    @GetMapping
+    public ResponseEntity<?> fetchMessages(@RequestParam(value = "page", defaultValue = "0") int page,
+                                           @RequestParam(value = "limit", defaultValue = "30") int limit,
+                                           @RequestParam(value = "order", defaultValue = "id") String order) {
+        ErrorMessage errorMessage;
+        List<Message> messageList = this.messageService.fetchMessages(page, limit, order);
+
+        if (messageList != null && messageList.size() > 0) {
+            List<MessageResponse> messageResponseList = messageList.stream().map(
+                    message -> new MessageResponse(message.getId(), message.getSubject(), message.getMessageContent(),
+                            message.getViews(), message.getOwner().getId(), message.getCreated(), message.getModified())
+            ).collect(Collectors.toList());
+
+            log.debug("Total records matching: {}", messageResponseList.size());
+            return new ResponseEntity<>(messageResponseList, HttpStatus.OK);
+        } else {
+            log.error("Did not find any record for request");
+            errorMessage = new ErrorMessage("Not found", new Date());
+        }
+
+        return new ResponseEntity<>(errorMessage, HttpStatus.NO_CONTENT);
+    }
+
+    @GetMapping(path = "/{messageId}")
+    public ResponseEntity<?> findMessage(@PathVariable Long messageId) {
+        ErrorMessage errorMessage;
+        Message message = this.messageService.findMessage(messageId);
+
+        if (message != null) {
+            MessageResponse messageResponse = new MessageResponse(message.getId(), message.getSubject(),
+                    message.getMessageContent(), message.getViews(), message.getOwner().getId(), message.getCreated(),
+                    message.getModified());
+            log.debug("Call to fetchMessages successful");
+            return new ResponseEntity<>(messageResponse, HttpStatus.OK);
+        } else {
+            log.error("Did not find any record with ID {}", messageId);
+            errorMessage = new ErrorMessage("Not found", new Date());
+        }
+        return new ResponseEntity<>(errorMessage, HttpStatus.NO_CONTENT);
     }
 }
